drive mechanics class, wrap atomics, size_t and uint tidy, kill some pico typedefs

+++ src/disc_image.cpp
@@ -1,6 +1,7 @@
 #include "disc_image.h"
 
 #include <ctype.h>
+#include <stddef.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
@@ -8,7 +9,6 @@
 #include "f_util.h"
 #include "ff.h"
 #include "logging.h"
-#include "pico/stdlib.h"
 #include "picostation.h"
 #include "subq.h"
 #include "third_party/iec-60908b/edcecc.h"
@@ -195,7 +195,7 @@ picostation::SubQ::Data picostation::DiscImage::generateSubQ(const int sector) {
         if (sector - c_leadIn < c_preGap) {
             m_currentLogicalTrack = 1;
         } else {
-            for (int i = 1; i < m_cueDisc.trackCount + 2; i++) {  // + 2 for lead in & lead out
+            for (size_t i = 1; i < m_cueDisc.trackCount + 2; i++) {  // + 2 for lead in & lead out
                 if (m_cueDisc.tracks[i + 1].indices[0] > sector - c_leadIn - c_preGap) {
                     m_currentLogicalTrack = i;
                     break;
@@ -238,7 +238,7 @@ picostation::SubQ::Data picostation::DiscImage::generateSubQ(const int sector) {
         case audioControlModes::NORMAL:
         case audioControlModes::ALTNORMAL:
         default:
-            for (int i = 0; i < 10; i++) {
+            for (size_t i = 0; i < 10; i++) {
                 subqdata.crc = (subqdata.crc << 8) ^ crc16_lut[((subqdata.crc >> 8) ^ subqdata.raw[i]) & 0xFF];
             }
             subqdata.crc = (subqdata.crc << 8) | (subqdata.crc >> 8);  // swap endianness
@@ -317,7 +317,7 @@ FRESULT picostation::DiscImage::load(const TCHAR *targetCue) {
 
     m_hasData = false;
     DEBUG_PRINT("Track\tStart\tLength\tPregap\n");
-    for (int i = 0; i <= m_cueDisc.trackCount + 1; i++) {
+    for (size_t i = 0; i <= m_cueDisc.trackCount + 1; i++) {
         if (m_cueDisc.tracks[i].trackType == CueTrackType::TRACK_TYPE_DATA) {
             m_hasData = true;
         }
@@ -332,7 +332,7 @@ void picostation::DiscImage::readData(void *buffer, const int sector) {
     FRESULT fr;
     UINT br = 0;
 
-    for (int i = 1; i <= m_cueDisc.trackCount + 1; i++) {
+    for (size_t i = 1; i <= m_cueDisc.trackCount + 1; i++) {
         if (sector < m_cueDisc.tracks[i + 1].indices[0]) {
             if (m_cueDisc.tracks[i].file->opaque) {
                 const int64_t seekBytes = (sector - m_cueDisc.tracks[i].fileOffset) * 2352LL;
